Conver to floats and group by hour not day.

+++ db/migrate/20220618125420_create_speedtests.rb
@@ -1,8 +1,8 @@
 class CreateSpeedtests < ActiveRecord::Migration[7.0]
   def change
     create_table :speedtests, id: :uuid do |t|
-      t.integer :upload
-      t.integer :download
+      t.float :upload
+      t.float :download
 
       t.timestamps
     end
